requirements.txt
djangorestframework==3.14.0
Django==4.2
django-cors-headers==3.14.0
drf-spectacular==0.26.2
gevent==22.10.2
gunicorn==20.1.0
mysqlclient==2.1.1
python-decouple==3.8
PyJWT==2.6.0
